makes social media optional
