+++ sucuri/database/dependency.py
@@ -1,0 +1,11 @@
+# IIB
+from sucuri.database.config import SessionLocal
+
+
+# Dependency
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
